+++ src/components/ImageLoader/index.js
@@ -1,4 +1,5 @@
 import ImgLoader from './ImgLoader';
+import Img from './Img';
 import ImageLoader from './ImageLoader';
 
-export { ImgLoader, ImageLoader };
+export { ImgLoader, ImageLoader, Img };