+++ src/components/ImageLoader/makeSrcSetCache.jsx
@@ -1,0 +1,6 @@
+let _cachedDefaultOptions;
+
+export const getCachedDefaultOptions = () => _cachedDefaultOptions;
+export const setCachedDefaultOptions = (options) => {
+  _cachedDefaultOptions = options;
+};