+++ src/components/index.js
@@ -6,7 +6,7 @@ import QuerystringSidebarWidget from '@kitconcept/volto-blocks/components/Widget
 import EffectiveDate from '@kitconcept/volto-blocks/components/EffectiveDate/EffectiveDate';
 import DisplayI18nDate from '@kitconcept/volto-blocks/components/DisplayI18nDate/DisplayI18nDate';
 
-import { ImgLoader, ImageLoader } from './ImageLoader';
+import { ImgLoader, ImageLoader, Img } from './ImageLoader';
 
 export {
   withStyleWrapper,
@@ -18,4 +18,5 @@ export {
   DisplayI18nDate,
   ImgLoader,
   ImageLoader,
+  Img,
 };